Completed all the features

+++ backend/controllers/goalController.js
@@ -27,7 +27,7 @@ const setGoal = async (req, res, next) => {
     }
     const goal = await Goal.create({
       text: req.body.text,
-      //   user: req.user.id,
+      user: req.user.id,
     });
     res.status(200).json(goal);
   } catch (err) {
@@ -48,19 +48,19 @@ const updateGoal = async (req, res, next) => {
       throw new Error("Goal not found");
     }
 
-    // if (!req.user) {
-    //   res.status(401);
-    //   throw new Error("User not found");
-    // }
+    if (!req.user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
 
-    // if (goal.user.toString() !== req.user.id) {
-    //   res.status(401);
-    //   throw new Error("User not authorized");
-    // }
+    if (goal.user.toString() !== req.user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
-    });
+    }).populate("user", "name");
 
     res.status(200).json(updatedGoal);
   } catch (err) {
@@ -81,19 +81,19 @@ const deleteGoal = async (req, res, next) => {
       throw new Error("Goal not found");
     }
 
-    // if (!req.user) {
-    //   res.status(401);
-    //   throw new Error("User not found");
-    // }
+    if (!req.user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
 
-    // if (goal.user.toString() !== req.user.id) {
-    //   res.status(401);
-    //   throw new Error("User not authorized");
-    // }
+    if (goal.user.toString() !== req.user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
 
-    const message = await Goal.deleteOne(goal);
+    await Goal.deleteOne(goal);
 
-    res.status(200).json(message);
+    res.status(200).json({ id: req.params.id });
   } catch (err) {
     next(err);
   }
